renamed package from com.rht_labs to com.redhat.labs
