+++ src/main/java/com/redhat/labs/omp/models/CreateProjectRequest.java
@@ -1,4 +1,4 @@
-package com.rht_labs.omp.models;
+package com.redhat.labs.omp.models;
 
 import javax.json.bind.annotation.JsonbProperty;
