+++ src/test/java/com/redhat/labs/ExampleResourceTest.java
@@ -1,4 +1,4 @@
-package com.rht_labs;
+package com.redhat.labs;
 
 import io.quarkus.test.junit.QuarkusTest;
 import org.junit.jupiter.api.Test;
