--- src/test/java/com/rht_labs/NativeExampleResourceIT.java
@@ -1,9 +1,0 @@
-package com.rht_labs;
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class NativeExampleResourceIT extends ExampleResourceTest {
-
-    // Execute the same tests but in native mode.
-}
